refactor(web-ui): decompose Analytics performance page
Split the 613-line Analytics page into a slim container plus
presentational pieces under components/domain/analytics/: pure stats
helpers and computeAnalyticsStats (analyticsStats.ts), the equity-curve
and R-distribution SVG charts (AnalyticsCharts.tsx), stat/insight cards
and the how-to-read explainer (AnalyticsCards.tsx), and the closed-trade
table (AnalyticsTradeTable.tsx). Container keeps the query and useMemo;
rendered DOM and public interface unchanged.
